# openspec/specifications/apis/openapi.yaml
openapi: 3.0.3
info:
  title: AI Tool Usage Tracker API
  version: 1.0.0
  description: |
    REST API for the AI Tool Usage Tracker platform (Phase 1 MVP).

    **Authentication:** JWT Bearer token except `/auth/login`, `/auth/refresh`, and `/health`.

    **Errors:** RFC 7807 Problem Details (`application/problem+json`).

    **Traceability:** FR-PLT-001, ADR-012, NFR-SEC-003/004/006.

    ## RBAC Roles and Access

    | Role | Scope |
    |---|---|
    | `super_admin` | Full access to all endpoints and data across the entire organisation |
    | `team_admin` | Access to members, tools (read), teams (read), thresholds, uploads, collectors (read), and dashboard analytics — all scoped to managed teams only |
    | `finance_viewer` | Access to reports (read) and `GET /dashboard/my-usage` only |
    | `auditor` | Access to audit logs and `GET /dashboard/my-usage` only |
    | `team_member` | Access to `GET /dashboard/my-usage` only |
  contact:
    name: Platform Team
  license:
    name: Proprietary
servers:
  - url: https://api.ai-tool-tracker.example/api/v1
    description: Production
  - url: https://staging-api.ai-tool-tracker.example/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /teams:
    get:
      tags: [Teams]
      summary: List teams
      description: '**Roles:** `super_admin`, `team_admin`.'
      operationId: listTeams
      parameters:
        - $ref: './components/parameters.yaml#/Limit'
        - $ref: './components/parameters.yaml#/Cursor'
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Team list
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/TeamListResponse'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
    post:
      tags: [Teams]
      summary: Create team
      description: '**Roles:** `super_admin`. **Validation:** unique name within org.'
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/TeamCreateRequest'
            examples:
              default:
                $ref: './components/examples.yaml#/TeamCreateExample'
      responses:
        '201':
          description: Team created
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/Team'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '409':
          $ref: './components/responses.yaml#/Conflict'
  /teams/{teamId}:
    get:
      tags: [Teams]
      summary: Get team by ID
      description: '**Roles:** `super_admin`, `team_admin`.'
      operationId: getTeam
      parameters:
        - $ref: './components/parameters.yaml#/TeamIdPath'
      responses:
        '200':
          description: Team details
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/Team'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '404':
          $ref: './components/responses.yaml#/NotFound'
    patch:
      tags: [Teams]
      summary: Update team
      description: '**Roles:** `super_admin`. Deactivate via `active: false`.'
      operationId: updateTeam
      parameters:
        - $ref: './components/parameters.yaml#/TeamIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/TeamUpdateRequest'
      responses:
        '200':
          description: Team updated
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/Team'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '404':
          $ref: './components/responses.yaml#/NotFound'
  /teams/{teamId}/members:
    get:
      tags: [Teams]
      summary: List team members
      description: '**Roles:** `super_admin`, `team_admin`.'
      operationId: listTeamMembers
      parameters:
        - $ref: './components/parameters.yaml#/TeamIdPath'
        - $ref: './components/parameters.yaml#/Limit'
        - $ref: './components/parameters.yaml#/Cursor'
      responses:
        '200':
          description: Team members
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: './components/schemas.yaml#/TeamMember'
                  meta:
                    $ref: './components/schemas.yaml#/PaginationMeta'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '404':
          $ref: './components/responses.yaml#/NotFound'
    post:
      tags: [Teams]
      summary: Add team member
      description: '**Roles:** `super_admin`, `team_admin` (administered teams only).'
      operationId: addTeamMember
      parameters:
        - $ref: './components/parameters.yaml#/TeamIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/TeamMemberAddRequest'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/TeamMember'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '404':
          $ref: './components/responses.yaml#/NotFound'
        '409':
          $ref: './components/responses.yaml#/Conflict'
  /teams/{teamId}/members/{userId}:
    delete:
      tags: [Teams]
      summary: Remove team member
      description: Historical usage retained. **Roles:** `super_admin`, scoped `team_admin`.
      operationId: removeTeamMember
      parameters:
        - $ref: './components/parameters.yaml#/TeamIdPath'
        - $ref: './components/parameters.yaml#/MemberUserId'
      responses:
        '204':
          description: Member removed
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '404':
          $ref: './components/responses.yaml#/NotFound'
...
  /thresholds:
    get:
      tags: [Thresholds]
      summary: List thresholds
      operationId: listThresholds
      parameters:
        - $ref: './components/parameters.yaml#/TeamId'
        - $ref: './components/parameters.yaml#/ToolId'
        - $ref: './components/parameters.yaml#/Limit'
        - $ref: './components/parameters.yaml#/Cursor'
      responses:
        '200':
          description: Threshold list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: './components/schemas.yaml#/Threshold'
                  meta:
                    $ref: './components/schemas.yaml#/PaginationMeta'
        '401':
          $ref: './components/responses.yaml#/Unauthorized'
    post:
      tags: [Thresholds]
      summary: Create threshold
      description: |
        **Roles:** `super_admin`, `team_admin`.

        **team_admin:** `team_id` in the request body must belong to one of the caller's
        managed teams. Supplying a `team_id` outside that scope returns `403`.

        **Validation:** `limit_value` >= 0; scope-specific IDs required
        (e.g. `team_id` when scope is `team`).
      operationId: createThreshold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './components/schemas.yaml#/ThresholdCreateRequest'
            examples:
              default:
                $ref: './components/examples.yaml#/ThresholdCreateExample'
      responses:
        '201':
          description: Threshold created
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/Threshold'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '422':
          $ref: './components/responses.yaml#/UnprocessableEntity'
  /thresholds/{thresholdId}:
    patch:
      tags: [Thresholds]
      summary: Update threshold
      description: |
        **Roles:** `super_admin`, `team_admin`.

        **team_admin:** operation on a threshold outside managed teams returns `404`.
      operationId: updateThreshold
      parameters:
        - $ref: './components/parameters.yaml#/ThresholdId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                limit_value:
                  type: number
                  minimum: 0
                severity:
                  $ref: './components/schemas.yaml#/AlertSeverity'
                active:
                  type: boolean
                notify_email:
                  type: boolean
                notify_in_app:
                  type: boolean
      responses:
        '200':
          description: Threshold updated
          content:
            application/json:
              schema:
                $ref: './components/schemas.yaml#/Threshold'
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '404':
          $ref: './components/responses.yaml#/NotFound'
    delete:
      tags: [Thresholds]
      summary: Delete threshold
      description: |
        **Roles:** `super_admin`, `team_admin`.

        **team_admin:** operation on a threshold outside managed teams returns `404`.
      operationId: deleteThreshold
      parameters:
        - $ref: './components/parameters.yaml#/ThresholdId'
      responses:
        '204':
          description: Threshold deleted
        '403':
          $ref: './components/responses.yaml#/Forbidden'
        '404':
          $ref: './components/responses.yaml#/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from POST /auth/login. Required on all endpoints except /auth/login, /auth/refresh, /health.